+++ spec/rubocop/formatter/clang_style_formatter_spec.rb
@@ -64,6 +64,25 @@ module Rubocop
                                        '^^',
                                        ''].join("\n")
         end
+
+        it 'does not display offending source line if it is blank' do
+          cop = Cop::Cop.new
+          source_buffer = Parser::Source::Buffer.new('test', 1)
+          source_buffer.source = (['     ', 'yaba']).to_a.join($RS)
+          cop.add_offence(:convention,
+                          Parser::Source::Range.new(source_buffer, 0, 2),
+                          'message 1')
+          cop.add_offence(:fatal,
+                          Parser::Source::Range.new(source_buffer, 6, 4),
+                          'message 2')
+
+          formatter.report_file('test', cop.offences)
+          expect(output.string).to eq ['test:1:1: C: message 1',
+                                       'test:2:1: F: message 2',
+                                       'yaba',
+                                       '^^^^',
+                                       ''].join("\n")
+        end
       end
     end
   end